split out in-app manager
